+++ python/hopsworks/mcp/models/project.py
@@ -1,0 +1,33 @@
+#
+#   Copyright 2025 Hopsworks AB
+#
+#   Licensed under the Apache License, Version 2.0 (the "License");
+#   you may not use this file except in compliance with the License.
+#   You may obtain a copy of the License at
+#
+#       http://www.apache.org/licenses/LICENSE-2.0
+#
+#   Unless required by applicable law or agreed to in writing, software
+#   distributed under the License is distributed on an "AS IS" BASIS,
+#   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#   See the License for the specific language governing permissions and
+#   limitations under the License.
+#
+
+from pydantic import BaseModel
+
+
+class Project(BaseModel):
+    """Model representing a project in Hopsworks MCP."""
+
+    name: str
+    id: int
+    owner: str
+    description: str
+    created: str
+
+
+class Projects(BaseModel):
+    """Model representing a collection of projects in Hopsworks MCP."""
+
+    projects: list[Project]
