+++ python/hopsworks/mcp/prompts/__init__.py
