+++ python/hopsworks/mcp/prompts/project.py
