+++ python/hopsworks/mcp/resources/__init__.py
