+++ python/hopsworks/mcp/server.py
